Check for get/set pairs in ES5 object literals
Enforce style requirement that setters should immediately follow getters
of the same name when both are specified.  This matches the old behavior
when both are specified and seems to be a good stylistic practice.

Note:  I couldn't figure out why the token string was emptied for the
gettter and setter property name token.  Leaving it intact is necessary
for checking that the getter immediately preceded the setter and doesn't
appear to cause any problems... can rework if necessary.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>