Adding 'nofuncname' option
This allows for bypassing errors on functions that do not
have a name. A primary example of a need for this would be a
CouchDB JS Design Document, which commonly will look like this:

function (doc) {
    ...
}

This feature will allow authors of CouchDB Design Documents
to validate their documents with JSLint, which they are currently
unable to do easily.